Now creates new directories if given exists

+++ _symmetry.py
@@ -13,8 +13,9 @@ import spglib
 
 if len(argv) < 3:
     print('Feed me with:')
-    print('(i)  name of  input directory')
-    print('(ii) name of output directory')
+    print('(i)   name of  input directory')
+    print('(ii)  name of output directory')
+    print('(iii) should the 1.P1 cells be deleted') 
     exit(-1)
 
 sameSymmetries = {}
@@ -22,7 +23,7 @@ sameSymmetries = {}
 directory = os.fsencode(argv[1])
 cntr = counter.Counter(limit=len(os.listdir(directory)),PREFIX='Processed ',SUFFIX=' files.')
 cntr.initiate()
-if len(argv) >= 4:
+if len(argv) >= 4 and 'n' not in argv[3] and 'N' not in argv[3]:
     cntr2 = counter.Counter(PREFIX='Removed   ',SUFFIX=' 1.P1 files.')
     cntr2.initiate()
     cntr.iterate_position()
@@ -39,7 +40,7 @@ for file in os.listdir(directory):
         print(filename, end='  ')
         print('('+argv[1]+'/'+filename+')')
         cntr.iterate_position()
-        if len(argv) >= 4:
+        if len(argv) >= 4 and 'n' not in argv[3] and 'N' not in argv[3]:
             cntr2.iterate_position()
         continue
     stdCell = spglib.standardize_cell(loader(0)['cellSymmetry'], to_primitive=1, no_idealize=0, symprec=1e-2)
@@ -47,7 +48,7 @@ for file in os.listdir(directory):
     local_key = str(o['number'])+'_'+o['international']
     local_key = re.sub('[^-0-9a-zA-Z]','.',local_key)
     
-    if len(argv) >= 4:
+    if len(argv) >= 4 and 'n' not in argv[3] and 'N' not in argv[3]:
         if local_key == '1.P1':
             os.remove(argv[1]+'/'+filename)
             cntr2.update()
@@ -60,17 +61,26 @@ for file in os.listdir(directory):
         sameSymmetries[local_key]  = [filename]
         print('Found new symmetry:',local_key, "(equivalent atoms)")
         cntr.iterate_position()
-        if len(argv) >= 4:
+        if len(argv) >= 4 and 'n' not in argv[3] and 'N' not in argv[3]:
             cntr2.iterate_position()
     else:
         sameSymmetries[local_key].append(filename)
     cntr.update()
 
 #print('All symmetries:')
-os.mkdir(argv[2])
+try:
+    os.mkdir(argv[2])
+    dirname=argv[2]
+except FileExistsError as err:
+    print(err,end='. ')
+    dirname=argv[2]+'.'+str(np.random.randint(int(1e9)))
+    print('Using '+dirname+' instead.')
+    with open('different_directory.txt','w+') as tmpfile:
+        tmpfile.write(dirname)
+    os.mkdir(dirname)
+
 for k in sameSymmetries.keys():
-    print(k)
-    name=argv[2]+'/'+k
+    name=dirname+'/'+k
     os.mkdir(name)
     for f in sameSymmetries[k]:
         shutil.copy2(argv[1]+'/'+f,name+'/')
